feat: implement dynamic messaging and user dashboard
Introduces dynamic slide messaging based on whether a wrapped presentation is personal or server-wide. This allows for personalized language like "Your" and "You" for individual users, and collective or server-named language like "Our", "We", or "{ServerName}'s" for server-wide presentations. This enhances the user experience by making the content more relevant to the viewer.

Also implements a new `/dashboard` route for non-admin users, providing them with a dedicated page to access their personal wrapped and the server-wide wrapped. Admins are redirected to `/admin`, and non-authenticated users are redirected to the homepage. This improves navigation and clarifies access paths for different user roles.

The server's friendly name is now fetched from Plex and cached to support server-specific messaging.

BREAKING CHANGE: The main landing page `/` now redirects authenticated users to either `/admin` or `/dashboard`. Previously, it redirected to the current year's wrapped. Updated the main page logic to align with new user dashboards.